Refactor gradientBorder into modifier extension
