+++ app/src/main/java/com/appdev/all_in_android/ui/components/bet_tracker/BetTrackerInfoCard.kt
@@ -1,7 +1,6 @@
 package com.appdev.all_in_android.ui.components.bet_tracker
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.height
@@ -11,30 +10,23 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.appdev.all_in_android.ui.theme.gradientBorder
 
 @Composable
 fun BetTrackerInfoCard(
     titleText: String,
     bodyText: String
 ) {
-    val gradientBrush = Brush.linearGradient(
-        colors = listOf(Color(0xFF1F70C7), Color(0xFF7DF3FE), Color(0xFF887DFE), Color(0xFF7D97FE))
-    )
     Column(
         modifier = Modifier
             .width(164.dp)
             .height(84.dp)
-            .border(
-                width = 1.dp,
-                brush = gradientBrush,
-                shape = RoundedCornerShape(8.dp)
-            )
+            .gradientBorder()
             .background(
                 color = Color(0xFF201E2D),
                 shape = RoundedCornerShape(8.dp)
