+++ app/src/main/java/com/appdev/all_in_android/ui/components/general/BetCard.kt
@@ -16,13 +16,13 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil3.compose.AsyncImage
+import com.appdev.all_in_android.ui.theme.gradientBorder
 
 @Composable
 fun BetCard(
@@ -34,18 +34,11 @@ fun BetCard(
     gain: Double,
     playerImageUrl: String = "",
     onClick: () -> Unit
-){
-    val gradientBrush = Brush.linearGradient(
-        colors = listOf(Color(0xFF1F70C7), Color(0xFF7DF3FE), Color(0xFF887DFE), Color(0xFF7D97FE))
-    )
+) {
     Row(
         modifier = Modifier
             .fillMaxWidth()
-            .border(
-                width = 1.dp,
-                brush = gradientBrush,
-                shape = RoundedCornerShape(8.dp)
-            )
+            .gradientBorder()
             .background(
                 color = Color(0xFF201E2D),
                 shape = RoundedCornerShape(8.dp)
@@ -56,8 +49,8 @@ fun BetCard(
             ),
         verticalAlignment = Alignment.CenterVertically,
         horizontalArrangement = Arrangement.SpaceBetween
-    ){
-        Row(){
+    ) {
+        Row() {
             AsyncImage(
                 model = playerImageUrl,
                 contentDescription = "player image",
@@ -127,7 +120,7 @@ private fun CostGainTextBody(cost: Double, gain: Double) {
 
 @Preview(showBackground = true)
 @Composable
-private fun BetCardPreview(){
+private fun BetCardPreview() {
     BetCard(
         title = "Jake Shane v. Harvard",
         date = "03/24",
@@ -135,5 +128,5 @@ private fun BetCardPreview(){
         bettingLine = "Scores first goal of game",
         cost = 20.00,
         gain = 40.00
-    ){}
+    ) {}
 }
